[APM][ECO] Removing Entities inventory from APM (#195116)
closes https://github.com/elastic/kibana/issues/194114

We will no longer show the entity inventory page on the APM UI. So, I'm
removing all its code.

(cherry picked from commit c44b7de7a23fbd5b11b31cc8d39c5baffb6c8d6f)
