+++ x-pack/plugins/observability_solution/apm/kibana.jsonc
@@ -30,8 +30,7 @@
       "lens",
       "maps",
       "uiActions",
-      "logsDataAccess",
-      "entityManager"
+      "logsDataAccess"
     ],
     "optionalPlugins": [
       "actions",
