+++ x-pack/plugins/observability_solution/apm/public/components/routing/app_root/apm_header_action_menu/add_data_context_menu.tsx
@@ -22,8 +22,6 @@ import {
   collectServiceLogs,
   addApmData,
 } from '../../../shared/add_data_buttons/buttons';
-import { ServiceEcoTour } from '../../../shared/entity_enablement/service_eco_tour';
-import { useEntityManagerEnablementContext } from '../../../../context/entity_manager_context/use_entity_manager_enablement_context';
 
 const addData = i18n.translate('xpack.apm.addDataContextMenu.link', {
   defaultMessage: 'Add data',
@@ -31,7 +29,6 @@ const addData = i18n.translate('xpack.apm.addDataContextMenu.link', {
 
 export function AddDataContextMenu() {
   const [popoverOpen, setPopoverOpen] = useState(false);
-  const { tourState, updateTourState } = useEntityManagerEnablementContext();
   const { services } = useKibana<ApmPluginStartDeps & ApmServices>();
   const {
     core: {
@@ -93,23 +90,17 @@ export function AddDataContextMenu() {
     },
   ];
 
-  const handleTourClose = () => {
-    updateTourState({ isTourActive: false });
-    setPopoverOpen(false);
-  };
   return (
     <>
       <EuiPopover
         id="integrations-menu"
         button={button}
-        isOpen={popoverOpen || tourState.isTourActive}
+        isOpen={popoverOpen}
         closePopover={() => setPopoverOpen(false)}
         panelPaddingSize="none"
         anchorPosition="downRight"
       >
-        <ServiceEcoTour onFinish={handleTourClose}>
-          <EuiContextMenu initialPanelId={0} panels={panels} />
-        </ServiceEcoTour>
+        <EuiContextMenu initialPanelId={0} panels={panels} />
       </EuiPopover>
     </>
   );
