+++ x-pack/plugins/observability_solution/apm/public/components/routing/app_root/index.tsx
@@ -42,7 +42,6 @@ import { RedirectWithDefaultEnvironment } from './redirect_with_default_environm
 import { RedirectWithOffset } from './redirect_with_offset';
 import { ScrollToTopOnPathChange } from './scroll_to_top_on_path_change';
 import { UpdateExecutionContextOnRouteChange } from './update_execution_context_on_route_change';
-import { EntityManagerEnablementContextProvider } from '../../../context/entity_manager_context/entity_manager_context';
 
 const storage = new Storage(localStorage);
 
@@ -83,17 +82,15 @@ export function ApmAppRoot({
                                     <BreadcrumbsContextProvider>
                                       <UrlParamsProvider>
                                         <LicenseProvider>
-                                          <EntityManagerEnablementContextProvider>
-                                            <AnomalyDetectionJobsContextProvider>
-                                              <InspectorContextProvider>
-                                                <ApmThemeProvider>
-                                                  <MountApmHeaderActionMenu />
-                                                  <Route component={ScrollToTopOnPathChange} />
-                                                  <RouteRenderer />
-                                                </ApmThemeProvider>
-                                              </InspectorContextProvider>
-                                            </AnomalyDetectionJobsContextProvider>
-                                          </EntityManagerEnablementContextProvider>
+                                          <AnomalyDetectionJobsContextProvider>
+                                            <InspectorContextProvider>
+                                              <ApmThemeProvider>
+                                                <MountApmHeaderActionMenu />
+                                                <Route component={ScrollToTopOnPathChange} />
+                                                <RouteRenderer />
+                                              </ApmThemeProvider>
+                                            </InspectorContextProvider>
+                                          </AnomalyDetectionJobsContextProvider>
                                         </LicenseProvider>
                                       </UrlParamsProvider>
                                     </BreadcrumbsContextProvider>
