--- x-pack/plugins/observability_solution/apm/public/context/entity_manager_context/use_entity_manager_enablement_context.ts
@@ -1,13 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { useContext } from 'react';
-import { EntityManagerEnablementContext } from './entity_manager_context';
-
-export function useEntityManagerEnablementContext() {
-  return useContext(EntityManagerEnablementContext);
-}
