+++ x-pack/plugins/observability_solution/apm/public/services/telemetry/telemetry_client.ts
@@ -9,9 +9,7 @@ import { AnalyticsServiceSetup } from '@kbn/core-analytics-browser';
 import {
   ITelemetryClient,
   SearchQuerySubmittedParams,
-  EntityExperienceStatusParams,
   TelemetryEventTypes,
-  EntityInventoryPageStateParams,
   EntityInventoryAddDataParams,
   EmptyStateClickParams,
 } from './types';
@@ -29,14 +27,6 @@ export class TelemetryClient implements ITelemetryClient {
       timerange,
       action,
     });
-  };
-
-  public reportEntityExperienceStatusChange = (params: EntityExperienceStatusParams) => {
-    this.analytics.reportEvent(TelemetryEventTypes.ENTITY_EXPERIENCE_STATUS, params);
-  };
-
-  public reportEntityInventoryPageState = (params: EntityInventoryPageStateParams) => {
-    this.analytics.reportEvent(TelemetryEventTypes.ENTITY_INVENTORY_PAGE_STATE, params);
   };
 
   public reportEntityInventoryAddData = (params: EntityInventoryAddDataParams) => {
