+++ x-pack/plugins/observability_solution/apm/public/services/telemetry/types.ts
@@ -21,14 +21,6 @@ export interface SearchQuerySubmittedParams {
   action: SearchQueryActions;
 }
 
-export interface EntityExperienceStatusParams {
-  status: 'enabled' | 'disabled';
-}
-
-export interface EntityInventoryPageStateParams {
-  state: 'empty_state' | 'available';
-}
-
 export interface EntityInventoryAddDataParams {
   view: 'empty_state' | 'add_data_button' | 'add_apm_cta' | 'add_apm_n/a';
   journey?: 'add_apm_agent' | 'associate_existing_service_logs' | 'collect_new_service_logs';
@@ -40,15 +32,11 @@ export interface EmptyStateClickParams {
 
 export type TelemetryEventParams =
   | SearchQuerySubmittedParams
-  | EntityExperienceStatusParams
-  | EntityInventoryPageStateParams
   | EntityInventoryAddDataParams
   | EmptyStateClickParams;
 
 export interface ITelemetryClient {
   reportSearchQuerySubmitted(params: SearchQuerySubmittedParams): void;
-  reportEntityExperienceStatusChange(params: EntityExperienceStatusParams): void;
-  reportEntityInventoryPageState(params: EntityInventoryPageStateParams): void;
   reportEntityInventoryAddData(params: EntityInventoryAddDataParams): void;
   reportTryItClick(params: EmptyStateClickParams): void;
   reportLearnMoreClick(params: EmptyStateClickParams): void;
@@ -56,8 +44,6 @@ export interface ITelemetryClient {
 
 export enum TelemetryEventTypes {
   SEARCH_QUERY_SUBMITTED = 'Search Query Submitted',
-  ENTITY_EXPERIENCE_STATUS = 'entity_experience_status',
-  ENTITY_INVENTORY_PAGE_STATE = 'entity_inventory_page_state',
   ENTITY_INVENTORY_ADD_DATA = 'entity_inventory_add_data',
   TRY_IT_CLICK = 'try_it_click',
   LEARN_MORE_CLICK = 'learn_more_click',
