+++ x-pack/plugins/observability_solution/apm/tsconfig.json
@@ -119,7 +119,6 @@
     "@kbn/react-kibana-context-theme",
     "@kbn/test-jest-helpers",
     "@kbn/security-plugin-types-common",
-    "@kbn/entityManager-plugin",
     "@kbn/server-route-repository-utils",
     "@kbn/core-analytics-browser",
     "@kbn/apm-types",
